requirements.txt
django-adminlte3
django-baton
django-filter
django-jet
djangorestframework
asgiref
certifi
chardet
Django
django-ckeditor
django-js-asset
django-modeltranslation
django-rosetta
idna
Pillow
polib
psycopg2-binary
pytz
requests
six
sqlparse
typing-extensions
Unidecode
urllib3
xmltodict
uwsgi
celery
django-celery-results
redis
bs4